Moved common sources to PlayerControls
