+++ entry_types/scrolled/package/spec/frontend/PlayerControls/ClassicPlayerControls/ProgressIndicators-spec.js
