+++ entry_types/scrolled/package/spec/frontend/PlayerControls/PlayPauseButton-spec.js
@@ -2,7 +2,7 @@ import React from 'react';
 import '@testing-library/jest-dom/extend-expect'
 import {render} from '@testing-library/react'
 
-import {PlayPauseButton} from 'frontend/PlayerControls/ClassicPlayerControls/PlayPauseButton';
+import {PlayPauseButton} from 'frontend/PlayerControls/PlayPauseButton';
 import {useFakeTranslations} from 'pageflow/testHelpers';
 
 describe('PlayPauseButton', () => {
