+++ entry_types/scrolled/package/src/frontend/PlayerControls/ControlBar.module.css
@@ -26,7 +26,7 @@
 }
 
 .button {
-  composes: unstyledButton from '../../utils.module.css';
+  composes: unstyledButton from '../utils.module.css';
   width: 32px;
   height: 32px;
   margin-right: 4px;
