+++ entry_types/scrolled/package/src/frontend/PlayerControls/MenuBarButton.module.css
@@ -54,7 +54,7 @@
 }
 
 .subMenuItemButton {
-  composes: unstyledButton from '../../utils.module.css';
+  composes: unstyledButton from '../utils.module.css';
   width: 100%;
   padding: 5px 10px 5px 25px;
   color: #fff;
