+++ entry_types/scrolled/package/src/frontend/PlayerControls/PlayPauseButton.js
@@ -3,7 +3,7 @@ import controlBarStyles from "./ControlBar.module.css";
 import PlayIcon from "./images/play.svg";
 import PauseIcon from "./images/pause.svg";
 
-import {useI18n} from '../../i18n';
+import {useI18n} from '../i18n';
 
 export function PlayPauseButton(props) {
   const {t} = useI18n();
